feat: move header to separate module
