+++ src/app/app.module.ts
@@ -3,38 +3,24 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { FormsModule } from '@angular/forms';
-
-import { MatIconModule } from '@angular/material/icon';
-import { MatCardModule } from '@angular/material/card';
-import { MatButtonModule } from '@angular/material/button';
 
 import { AppComponent } from './app.component';
-import { UserBlockComponent } from './header/user-block/user-block.component';
-import { HeaderComponent } from './header/header/header.component';
-import { InputBlockComponent } from './header/input-block/input-block.component';
-import { SettingsComponent } from './header/settings/settings.component';
-import { FilteringComponent } from './header/filtering/filtering.component';
+
 import { ResultsListComponent } from './search-panel/results-list/results-list.component';
 import { ResultItemComponent } from './search-panel/result-item/result-item.component';
-import { FooterDirective } from './search-panel/result-item/footer.directive';
 import {
   FilterByDate,
   FilterByViews,
   FilterByWordPipe,
 } from './search-panel/results-list/filtering.pippe';
 import { HttpClientModule } from '@angular/common/http';
-import { ShowListService } from './services/show-list.service';
-import { SortService } from './services/search-sort.service';
+import { MaterialModule } from './material.module';
+import { CoreModule } from './core/core.module';
+import { FooterDirective } from './search-panel/result-item/footer.directive';
 
 @NgModule({
   declarations: [
     AppComponent,
-    UserBlockComponent,
-    HeaderComponent,
-    InputBlockComponent,
-    SettingsComponent,
-    FilteringComponent,
     ResultsListComponent,
     ResultItemComponent,
     FooterDirective,
@@ -46,13 +32,11 @@ import { SortService } from './services/search-sort.service';
     BrowserModule,
     AppRoutingModule,
     BrowserAnimationsModule,
-    MatCardModule,
-    MatIconModule,
-    FormsModule,
-    MatButtonModule,
     HttpClientModule,
+    MaterialModule,
+    CoreModule,
   ],
-  providers: [SortService, ShowListService],
+  providers: [],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
