+++ src/app/core/components/header/input-block/input-block.component.html
