+++ src/app/core/components/header/settings/settings.component.ts
