+++ src/app/core/components/header/user-block/user-block.component.scss
