+++ src/app/search-panel/models/item.model.ts
