+++ src/app/search-panel/models/list-response.model.ts
