+++ src/app/search-panel/result-item/result-item.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { ItemType } from '../item.model';
+import { ItemType } from '../models/item.model';
 
 @Component({
   selector: 'app-result-item',
